Introduce stringifyObject utility function.
