+++ assets/js/googlesitekit/api/index.js
@@ -20,7 +20,6 @@
  * External dependencies
  */
 import invariant from 'invariant';
-import md5 from 'md5';
 
 /**
  * WordPress dependencies
@@ -37,7 +36,7 @@ import {
 	getKeys,
 	setItem,
 } from 'assets/js/googlesitekit/api/cache';
-import { sortObjectProperties } from 'assets/js/util';
+import { stringifyObject } from 'assets/js/util';
 
 // Caching is enabled by default.
 let cachingEnabled = true;
@@ -67,7 +66,7 @@ export const createCacheKey = ( type, identifier, datapoint, queryParams = {} ) 
 		Object.keys( queryParams ).length
 	) {
 		keySections.push(
-			md5( JSON.stringify( sortObjectProperties( queryParams ) ) )
+			stringifyObject( queryParams )
 		);
 	}
