+++ assets/js/util/index.js
@@ -54,6 +54,7 @@ import { tagMatchers as tagmanagerTagMatchers } from '../modules/tagmanager/util
 import { trackEvent } from './tracking';
 export { trackEvent };
 export * from './sanitize';
+export * from './stringify';
 export * from './standalone';
 export * from './storage';
 export * from './i18n';
@@ -683,27 +684,6 @@ export function moduleIcon( module, blockedByParentModule, width = '33', height 
 	}
 
 	return iconComponent;
-}
-
-/**
- * Sorts an object by its keys.
- *
- * The returned value will be a sorted copy of the input object.
- * Any inner objects will also be sorted recursively.
- *
- * @param {Object} obj The data object to sort.
- * @return {Object} The sorted data object.
- */
-export function sortObjectProperties( obj ) {
-	const orderedData = {};
-	Object.keys( obj ).sort().forEach( ( key ) => {
-		let val = obj[ key ];
-		if ( val && 'object' === typeof val && ! Array.isArray( val ) ) {
-			val = sortObjectProperties( val );
-		}
-		orderedData[ key ] = val;
-	} );
-	return orderedData;
 }
 
 /**
